+++ aws-s3/src/it/scala/busymachines/pureharm/aws/s3/S3LiveTest.scala
@@ -32,18 +32,18 @@ import busymachines.pureharm.testkit._
   * @author Lorand Szakacs, https://github.com/lorandszakacs
   * @since 22 May 2019
   */
-final class S3LiveTest extends PureharmTestWithResource {
+final class S3LiveTest extends PureharmTest {
   implicit override val testLogger: TestLogger = TestLogger(Slf4jLogger.getLogger[IO])
-  val l = testLogger
+  private val l = testLogger
 
   private val UTF_8 = java.nio.charset.StandardCharsets.UTF_8
 
-  override type ResourceType = (S3Config, AmazonS3Client[IO])
-
-  override def resource(meta: MetaData): Resource[IO, ResourceType] = for {
-    config   <- S3Config.fromNamespaceR[IO]("test-live.pureharm.aws.s3")
-    s3Client <- AmazonS3Client.resource[IO](config)
-  } yield (config, s3Client)
+  private val resource = ResourceFixture[(S3Config, AmazonS3Client[IO])] { _ =>
+    for {
+      config   <- S3Config.fromNamespaceR[IO]("test-live.pureharm.aws.s3")
+      s3Client <- AmazonS3Client.resource[IO](config)
+    } yield (config, s3Client)
+  }
 
   private val f1S3Key: S3FileKey =
     S3FileKey[Try]("folder", "subfolder", "file.txt").get
@@ -57,7 +57,7 @@ final class S3LiveTest extends PureharmTestWithResource {
 
   private val f1_content_length: S3ContentLengthBytes = S3ContentLengthBytes(22L)
 
-  test("s3 init bucket + list bucket") { case (config, client) =>
+  resource.test("s3 init bucket + list bucket") { case (config, client) =>
     val newBucket = S3Bucket("testinitbucket")
     for {
       buckets <- client.listBuckets
@@ -67,10 +67,10 @@ final class S3LiveTest extends PureharmTestWithResource {
       _ <- Resource.make(client.initBucket(newBucket))(_ => client.deleteBucket(newBucket).attempt.void).use { _ =>
         client.listBuckets.map(bs => IO(assert(bs.contains(newBucket))))
       }
-    } yield succeed
+    } yield ()
   }
 
-  test("s3 put + get + delete") { case (config, client) =>
+  resource.test("s3 put + get + delete") { case (config, client) =>
     for {
       _   <- l.info(s"acquired client resource: ${client.toString}")
       _   <-
@@ -97,10 +97,10 @@ final class S3LiveTest extends PureharmTestWithResource {
           .flatMap(g => l.error(s"SHOULD HAVE DELETED, but got: ${new String(g, UTF_8)}"))
           .void
           .handleErrorWith(t => l.info(s"AFTER DELETE — expected failure, and got it: ${t.toString}"))
-    } yield succeed
+    } yield ()
   }
 
-  test("s3 putStream + getStream + delete") { case (config, client) =>
+  resource.test("s3 putStream + getStream + delete") { case (config, client) =>
     for {
       _   <- l.info(s"acquired client resource: ${client.toString}")
       _   <-
@@ -123,10 +123,10 @@ final class S3LiveTest extends PureharmTestWithResource {
           .flatMap(g => l.error(s"SHOULD HAVE DELETED, but got: ${new String(g, UTF_8)}"))
           .void
           .handleErrorWith(t => l.info(s"AFTER DELETE — expected failure, and got it: ${t.toString}"))
-    } yield succeed
+    } yield ()
   }
 
-  test("s3 copy + exists + list") { case (config, client) =>
+  resource.test("s3 copy + exists + list") { case (config, client) =>
     for {
       _             <- l.info(s"acquired client resource: ${client.toString}")
       _             <-
@@ -166,7 +166,7 @@ final class S3LiveTest extends PureharmTestWithResource {
         client
           .delete(config.bucket, f2S3Key)
           .handleErrorWith(e => l.error(e)(s"DELETE failed w/: $e"))
-    } yield succeed
+    } yield ()
   }
 
 }