+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## dependency upgrades
 - [AWS SDK v1](https://github.com/aws/aws-sdk-java/releases) `1.11.991`
 - [AWS SDK v2](https://github.com/aws/aws-sdk-java-v2/releases) `2.16.34`
+- [pureharm-core-sprout](https://github.com/busymachines/pureharm-core/releases) `0.2.0`
+- [pureharm-core-anomaly](https://github.com/busymachines/pureharm-core/releases) `0.2.0`
+- [pureharm-effects-cats](https://github.com/busymachines/pureharm-effects-cats/releases) `0.2.0`
+- [pureharm-config](https://github.com/busymachines/pureharm-config/releases) `0.2.0`
 
 ## aws-logs